// postponed/auth-user-notifications/functions/src/types.ts
/*
 * Copyright 2019 Google LLC
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *    https://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

import * as firebase from "firebase-admin";

type Device = {
  deviceId: string;
  fcmToken: string;
  name: string;
  os: string;
  type: "Android" | "iOS" | "Web";
};
export type Devices = { [deviceId: string]: Device };
export type Notification = {
  android?: firebase.messaging.AndroidConfig;
  apns?: firebase.messaging.ApnsConfig;
  data?: { [key: string]: string };
  notification?: firebase.messaging.Notification;
  userId?: string;
  userIds?: string[];
  webpush?: firebase.messaging.WebpushConfig;
};
export type TokenMap = {
  [token: string]: {
    deviceId: string;
    userId: string;
  };
};
export type UserDevices = {
  devices: Devices;
  userId: string;
};